requirements.txt
flask
sentence-transformers
faiss-cpu
requests
pypdf
